refactor(haip-17): refactor import and export
