+++ domain/account/ui/Layout/index.tsx
@@ -1,3 +1,1 @@
-import Layout from "./Layout";
-
-export default Layout;
+export * from "./Layout";
