+++ domain/account/ui/LoginForm/LoginForm.tsx
@@ -1,6 +1,6 @@
 import { Button, Icon, TextInput } from "@/domain/shared/ui";
 
-const LoginForm = () => (
+export const LoginForm = () => (
   <div className="flex min-h-[369px] min-w-[406px] max-w-[406px] flex-col gap-4 p-[18px]">
     <h2 className="text-[32px] font-normal leading-[38px]">Masuk</h2>
     <TextInput label="Email" placeholder="Masukan email kamu" required />
