+++ domain/account/ui/LoginForm/index.ts
@@ -1,3 +1,1 @@
-import LoginForm from "./LoginForm";
-
-export default LoginForm;
+export * from "./LoginForm";
