+++ domain/account/ui/index.ts
@@ -1,4 +1,2 @@
-import LoginForm from "./LoginForm";
-import Layout from "./Layout";
-
-export { LoginForm, Layout };
+export * from "./LoginForm";
+export * from "./Layout";
