+++ domain/shared/configs/index.ts
@@ -1,0 +1,2 @@
+export * from "./muiThemeProvider";
+export * from "./theme";
