+++ domain/shared/configs/muiThemeProvider.tsx
@@ -2,9 +2,13 @@
 
 import { useEffect } from "react";
 import { ThemeProvider } from "@mui/material/styles";
-import theme from "./theme";
+import { theme } from "./theme";
 
-const MuiThemeProvider = ({ children }: { children: React.ReactNode }) => {
+export const MuiThemeProvider = ({
+  children,
+}: {
+  children: React.ReactNode;
+}) => {
   useEffect(() => {
     const rootElement = document.getElementById("__next");
     theme.components!.MuiPopover!.defaultProps!.container = rootElement;
