+++ domain/shared/ui/Button/index.ts
@@ -1,3 +1,1 @@
-import Button from "./Button";
-
-export default Button;
+export * from "./Button";
