+++ domain/shared/ui/Icon/index.ts
@@ -1,3 +1,1 @@
-import Icon from "./Icon";
-
-export default Icon;
+export * from "./Icon";
