+++ domain/shared/ui/TextInput/index.ts
@@ -1,3 +1,1 @@
-import TextInput from "./TextInput";
-
-export default TextInput;
+export * from "./TextInput";
